move app and db to utils
