+++ src/tests/after.test.ts
@@ -1,7 +1,7 @@
 import { Container } from 'typedi';
 import { Connection } from 'typeorm';
 
-import { Database } from '../libs/database';
+import { Database } from '../utils/database';
 
 after(async () => {
   const database = Container.get(Database);
