+++ src/tests/before.test.ts
@@ -1,7 +1,7 @@
 import { Container } from 'typedi/Container';
 import { Connection } from 'typeorm';
 
-import { Database } from '../libs/database';
+import { Database } from '../utils/database';
 
 before(async () => {
   const database = Container.get(Database);
