+++ src/tests/user/loginUser.controller.test.ts
@@ -5,7 +5,7 @@ import { Container } from 'typedi';
 import { Connection } from 'typeorm';
 
 import { IRole } from '../../enums';
-import { App } from '../../libs/app';
+import { App } from '../../utils/app';
 
 const app = Container.get(App).getApp();
 const server = supertest(app);
